crud province

+++ API/connect/src/main/java/com/learn/first/restapi/address/repository/IProvinceRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.learn.first.restapi.address.model.CProvince;
 
-public interface IProvinceRepository extends JpaRepository<CProvince, Long> {
+public interface IProvinceRepository extends JpaRepository<CProvince, Integer> {
     CProvince findByCode(String provinceCode);
 }
